requirements.txt
# Plant Disease Detection System - Dependencies
# requirements.txt

# Core ML and Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scikit-image==0.21.0

# Computer Vision
opencv-python==4.8.0.76
Pillow==10.0.0

# Web Framework
streamlit==1.25.0

# Plotting and Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Additional utilities
pickle-mixin==1.0.2
warnings